+++ symbolobj.c
@@ -132,7 +132,7 @@ static MUST_CHECK Obj *hash(Obj *o1, int *hs, linepos_t UNUSED(epoint)) {
     h = (unsigned int)*s.data << 7;
     l = s.len;
     while ((l--) != 0) h = (1000003 * h) ^ *s.data++;
-    h ^= s.len;
+    h ^= (unsigned int)s.len;
     *hs = v1->hash = h & ((~0U) >> 1);
     return NULL;
 }
